requirements.txt
discord.py
aiohttp
nest_asyncio
matplotlib
